refactor: 기물들 추상클래스 Piece로 추상화 (King, Queen, Bishop, Rook) 및 King 체크 위치로 이동 불가 조건 삭제
